[Lab 0]
>What happens if you annotate line 7 with a type annotation?
It allows the type of x to be inferred.
>What if you give x a value?
The code works fine. This time, it infers the type of x by using the value that was given.
>What if you do both?
It will use the type that was annotated but will check against the value. If the value is one type and the annotation is a different type, the compiler will tell you what is a valid and expected value.
>What type should x be, anyway?
Should be an integer type that's big enough to hold 10.
>What if x is the same type as 10? What if it's a different type?
If x is the same type as 10, then the comparison on line 8 is compiled properly. If not, the compiler will complain about a comparison between different types.